requirements.txt
# Core dependencies
python-dotenv==1.0.0

# Vector store
qdrant-client==1.7.3

# Embeddings
google-generativeai==0.3.2
sentence-transformers==2.3.1
torch

# Document parsing
PyPDF2==3.0.1
python-docx==1.1.0

# Utilities
tqdm==4.66.1
numpy
rank-bm25

streamlit==1.31.0